move element-wise functions outside
