+++ ggml/src/ggml-sycl/backend.hpp
@@ -28,5 +28,6 @@
 #include "im2col.hpp"
 #include "wkv6.hpp"
 #include "outprod.hpp"
+#include "element_wise.hpp"
 
 #endif // GGML_SYCL_BACKEND_HPP
